refactor: format

+++ src/module-system/a-authsimple/backend/src/bean/event.accountMigration.js
@@ -1,7 +1,6 @@
 module.exports = ctx => {
   const moduleInfo = ctx.app.meta.mockUtil.parseInfoFromPackage(__dirname);
   class eventBean {
-
     async execute(context, next) {
       const data = context.data;
       const modelAuthSimple = ctx.model.module(moduleInfo.relativeName).authSimple;
@@ -9,20 +8,13 @@ module.exports = ctx => {
       const authSimple = await modelAuthSimple.get({ userId: data.userIdFrom });
       if (authSimple) {
         // delete old record
-        await ctx.model.query(
-          'delete from aAuthSimple where deleted=0 and iid=? and userId=?',
-          [ctx.instance.id, data.userIdTo]
-        );
+        await ctx.model.query('delete from aAuthSimple where deleted=0 and iid=? and userId=?', [ctx.instance.id, data.userIdTo]);
         // update
-        await ctx.model.query(
-          'update aAuthSimple a set a.userId=? where a.deleted=0 and a.iid=? and a.userId=?',
-          [data.userIdTo, ctx.instance.id, data.userIdFrom]
-        );
+        await ctx.model.query('update aAuthSimple a set a.userId=? where a.deleted=0 and a.iid=? and a.userId=?', [data.userIdTo, ctx.instance.id, data.userIdFrom]);
       }
       // next
       await next();
     }
-
   }
 
   return eventBean;
